requirements.txt
pycryptodome
flask
flask-cors
requests
ecdsa